Dialog DeleteConfirm

+++ src/app/tasks/tasks-list/task-item/task-item.component.ts
@@ -4,6 +4,8 @@ import {ActivatedRoute, Router} from '@angular/router';
 import {TaskService} from '../../task.service';
 import {SharedService} from '../../../shared/shared.service';
 import {Subscription} from 'rxjs';
+import {MatDialog} from '@angular/material';
+import {DeleteConfirmComponent} from '../../../shared/delete-confirm.component';
 
 @Component({
   selector: 'app-task-item',
@@ -19,7 +21,8 @@ export class TaskItemComponent implements OnInit, OnDestroy {
   constructor(private router: Router,
               private route: ActivatedRoute,
               private taskService: TaskService,
-              private sharedService: SharedService) { }
+              private sharedService: SharedService,
+              private dialog: MatDialog) { }
 
   ngOnInit() {
     this.componentSubs.push(this.sharedService.isEmployeeModeChanged.subscribe(result => this.employeeMode = result));
@@ -34,7 +37,14 @@ export class TaskItemComponent implements OnInit, OnDestroy {
   }
 
   onDeleteTask(taskId: string) {
-    this.taskService.deleteTask(taskId);
+    const dialogRef = this.dialog.open(DeleteConfirmComponent);
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.taskService.deleteTask(taskId);
+      } else {
+        dialogRef.close();
+      }
+    });
   }
 
   ngOnDestroy(): void {
